Create app.py
In this coordinates converter Israeli-used geo-grids can be converted and displayed as links to map apps. 
A typo identifying feature is incorporated and an 'off-boundry' limiter to exclude out-of-Israel mistakes.
This app was hobbled together for Har-HaNegev Search and Rescue Team by The Couch in the South/Scorpion Man Exitprizes Ltd.